Add comment explaining tuple use.
This project originally started out as a means to test my app logic in a UWP project to see if I would have trouble doing it in WPF.
Needless to say, I didn't and before I knew it I ended up with a fully functional program that will be worked on instead for the time being.
Let's see where this goes.